Integrate SDL2 into the context
